Exercise (#53)
Can inject gpx files into the system and have them be added in the
correct location with a related markdown file.  This should at least get
the data recorded properly so that I can determine what functionality I
want to add to the reporting plugin after the fact.

Don't want to store stats in the markdown files at the moment since I
don't know what I want to put there.  Might as well allow the reporter
to calculate those values on it's own.

- Adding additional functionality for processing each entry.
- Add functionality to log_file to find files that are pointed to by the
  entry files.

In case it is the one that defines how the topic should be defined.  At
some point need to look up how to define the topics before all of the
subscriptions are put into place.

- Allow for copying of files as defined by reports
- Provide simple details about the gpx files that are defined

Closes #43